requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
duckdb
fastapi
h11
idna
pydantic
pydantic-core
python-dotenv
requests
sniffio
starlette
typing-extensions
typing-inspection
urllib3
uvicorn
watchdog